refactor(tree): Change TreeNode symbol type from u16 to u32 for improved compatibility

+++ runtime2/src/tree.rs
@@ -16,7 +16,7 @@ pub struct Tree {
 /// Internal tree node representation
 pub(crate) struct TreeNode {
     /// Symbol type
-    symbol: u16,
+    symbol: u32,
     /// Byte range in source
     start_byte: usize,
     end_byte: usize,
@@ -28,7 +28,7 @@ pub(crate) struct TreeNode {
 
 impl TreeNode {
     /// Create a new tree node with children
-    pub(crate) fn new_with_children(symbol: u16, start_byte: usize, end_byte: usize, children: Vec<TreeNode>) -> Self {
+    pub(crate) fn new_with_children(symbol: u32, start_byte: usize, end_byte: usize, children: Vec<TreeNode>) -> Self {
         Self {
             symbol,
             start_byte,
@@ -50,7 +50,7 @@ impl Tree {
     }
     
     /// Get the root node's kind
-    pub fn root_kind(&self) -> u16 {
+    pub fn root_kind(&self) -> u32 {
         self.root.symbol
     }
